% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSE_function.R
\name{assembleRSE}
\alias{assembleRSE}
\title{Collect data into SummarizedExperiment}
\usage{
assembleRSE(ctss, tcs, em, design)
}
\arguments{
\item{ctss}{GRangesList: CTSS-GRanges.}

\item{tcs}{GRanges: TCs}

\item{em}{matrix, integer: EM}

\item{design}{data.frame: Study note}
}
\value{
RangedSummarizedExperiment
}
\description{
Collects CTSS-GRanges, Tag Cluster (TCs), Expression Matrix (EM) and study design into a SummarizedExperiment. Additional options for documenting the object are available.
}
\note{
This function does not sort or match the input, so arguments must be provided with matching columns, rows, etc.
}
\examples{
# ADD_EXAMPLES_HERE
}

