% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancers.R
\name{balanceBC}
\alias{balanceBC}
\title{Balance statistic: Bhattacharyya coefficient (BC)}
\usage{
balanceBC(PD, MD, PU, MU)
}
\arguments{
\item{PD}{Plus-Downstream signal}

\item{MD}{Minus-Downstream signal}

\item{PU}{Plus-Upstream signal}

\item{MU}{Plus-Upstream signal}
}
\value{
Balance score of the same class as inputs.
}
\description{
Calculates the Bhattacharyya coefficient from observed plus/minus
upstream/downstream signals to a perfect bidirectional site, where
plus-downstream = 50% and minus-downstream = 50%
}
\examples{
# Unbalanced
balanceBC(2, 3, 1, 0)

# Balanced
balanceBC(3, 3, 0, 0)
}
