% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancers.R
\name{balanceD}
\alias{balanceD}
\title{Balance statistic: Andersson's D.}
\usage{
balanceD(PD, MD, PU, MU)
}
\arguments{
\item{PD}{Plus-Downstream signal}

\item{MD}{Minus-Downstream signal}

\item{PU}{Plus-Upstream signal}

\item{MU}{Plus-Upstream signal}
}
\value{
Balance score of the same class as inputs.
}
\description{
Calculates the D-statistics from Andersson et al the observed plus/minus
downstream signals. The D statistics is rescaled from -1:1 to 0:1 so it can
be used for slice-reduce identification of bidirectional sites.
}
\examples{
# Unbalanced
balanceD(2, 3, 1, 0)

# Balanced
balanceD(3, 3, 0, 0)
}
