% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_browser.R
\name{balanceTracks}
\alias{balanceTracks}
\title{Bidirectional balance Genome Browser Tracks}
\usage{
balanceTracks(ctssCoverage, window = 199, balanceFun = BC,
  plusColor = "cornflowerblue", minusColor = "tomato",
  balanceColor = "forestgreen", biocParallel = bpparam(), ...)
}
\arguments{
\item{ctssCoverage}{ctssCoverage GRanges: CTSS coverage.}

\item{window}{integer: Width of sliding window used for calculating sums.}

\item{balanceFun}{function: Function used to calculate balance (must have arguments plusUpstream, plusDownstream, minusUpstream, minusDownstream).}

\item{plusColor}{character: Color for plus-strand coverage.}

\item{minusColor}{character: Color for minus-strand coverage.}

\item{balanceColor}{character: Color for bidirectional balance.}

\item{biocParallel}{BiocParallelParam: Settings for parallel backend.}

\item{...}{additional arguments passed to balanceFun.}
}
\value{
list of 3 DataTracks for upstream, downstream and balance.
}
\description{
Visualize balance scores used for detectiong of bidirectional sites. Mainly intended as diagnostic tools for export users.
}
\note{
Potentially consumes a large amount of memory!
}
\examples{
# ADD_EXAMPLES_HERE
}
