% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bidirectional_clustering.R
\name{BC}
\alias{BC}
\title{Bhattacharyya coefficient.}
\usage{
BC(plusUpstream, plusDownstream, minusUpstream, minusDownstream)
}
\arguments{
\item{plusUpstream}{RleList or other: Sum of plus strand upstream (reference=0.0)}

\item{plusDownstream}{RleList or other: Sum of plus strand downstream (reference=0.5)}

\item{minusUpstream}{RleList or other: Sum of minus strand upstream (reference=0.0)}

\item{minusDownstream}{RleList or other: Sum of minus strand downstream (reference=0.5)}
}
\value{
Bhattacharyya coefficient
}
\description{
Quantifies departure from the reference bidirectional site (50%/50% of expression in bidirectional arms) using the Bhattacharyya coefficient as a measure of statistical distance.
}
\examples{
# ADD_EXAMPLES_HERE
}
