% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bidirectional_clustering.R
\name{bidirectionalClustering}
\alias{bidirectionalClustering}
\title{Bidirectional clustering of genome-wide coverage}
\usage{
bidirectionalClustering(ctssCoverage, window = 199, balanceThreshold = 0.99,
  balanceFun = BC, advancedStats = FALSE, biocParallel = bpparam(), ...)
}
\arguments{
\item{ctssCoverage}{GRanges: CTSS coverage.}

\item{window}{integer: Width of sliding window used for calculating sums.}

\item{balanceThreshold}{numeric: Minimum balance score to be considered a bidirectional site.}

\item{balanceFun}{function: Function used to calculate balance (must have arguments plusUpstream, plusDownstream, minusUpstream, minusDownstream).}

\item{advancedStats}{logical: Whether to return internal statistics for defining bidirectional sites.}

\item{biocParallel}{BiocParallelParam: Settings for parallel backend.}

\item{...}{additional arguments passed to balanceFun.}
}
\value{
GRanges with bidirectional sites: Minimum width is 1 + 2*window, TPM sum (on both strands) in the score column, and the maximal bidirectional site in the thick column. If advancedStats=TRUE, additional columns are added to the GRanges.
}
\description{
Finds sites with (balanced and divergent) bidirectional transcription using sliding windows of summed coverage. Custom functions for calculating balance can be used.
}
\examples{
# ADD_EXAMPLES_HERE
}
