% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.R
\name{bwCommonGenome}
\alias{bwCommonGenome}
\title{Find a common genome for a series of BigWig files.}
\usage{
bwCommonGenome(plusStrand, minusStrand, method = "intersect")
}
\arguments{
\item{plusStrand}{BigWigFileList: BigWig files with plus-strand CTSS data.}

\item{minusStrand}{BigWigFileList: BigWig files with minus-strand CTSS data.}

\item{method}{character: Either "intersect" or "union".}
}
\value{
Sorted Seqinfo-object.
}
\description{
Finds a common genome for a series of BigWig-files, either using only levels present in all files (intersect) or in any file (union).
}
\seealso{
Other BigWig functions: \code{\link{bwGenomeCompatibility}},
  \code{\link{bwValid}}
}
