% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.R
\name{bwGenomeCompatibility}
\alias{bwGenomeCompatibility}
\title{Check if BigWig-files are compatible with a given genome.}
\usage{
bwGenomeCompatibility(plusStrand, minusStrand, genome)
}
\arguments{
\item{plusStrand}{BigWigFileList: BigWig files with plus-strand CTSS data.}

\item{minusStrand}{BigWigFileList: BigWig files with minus-strand CTSS data.}

\item{genome}{Seqinfo: Genome information.}
}
\value{
TRUE, raises an error if the supplied genome is incompabtible.
}
\description{
Given a genome, checks whether a series of BigWig-files are compatible by checking if all common seqlevels have equal seqlengths.
}
\seealso{
Other BigWig functions: \code{\link{bwCommonGenome}},
  \code{\link{bwValid}}
}
