% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\docType{methods}
\name{calcShape}
\alias{calcShape}
\alias{calcShape,GRanges,GenomicRanges-method}
\alias{calcShape,RangedSummarizedExperiment,GenomicRanges-method}
\alias{calcShape,GRanges,RangedSummarizedExperiment-method}
\alias{calcShape,RangedSummarizedExperiment,RangedSummarizedExperiment-method}
\alias{calcShape,GRanges,GPos-method}
\title{Quantify Tag Cluster shapes}
\usage{
calcShape(object, pooled, ...)

\S4method{calcShape}{GRanges,GenomicRanges}(object, pooled,
  outputColumn = "IQR", shapeFunction = shapeIQR, ...)

\S4method{calcShape}{RangedSummarizedExperiment,GenomicRanges}(object, pooled,
  ...)

\S4method{calcShape}{GRanges,RangedSummarizedExperiment}(object, pooled, ...)


  \S4method{calcShape}{RangedSummarizedExperiment,RangedSummarizedExperiment}(object,
  pooled, ...)

\S4method{calcShape}{GRanges,GPos}(object, pooled, ...)
}
\arguments{
\item{object}{GenomicRanges or RangedSummarizedExperiment: Tag clusters.}

\item{pooled}{GenomicRanges or RangedSummarizedExperiment: Pooled CTSS as the score column.}

\item{...}{additional arguments passed to shapeFunction.}

\item{outputColumn}{character: Name of column to hold shape statistics.}

\item{shapeFunction}{function: Function to apply to each TC (See details).}
}
\value{
Adds a column
}
\description{
Apply a shape-function to the pooled CTSS of every TC.
}
\seealso{
Other Shape functions: \code{\link{shapeEntropy}},
  \code{\link{shapeIQR}}, \code{\link{shapeMultimodality}}
}
