% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calcSupport}
\alias{calcSupport}
\title{Calculate support of CAGE data.}
\usage{
calcSupport(object, inputAssay = "counts", outputColumn = "support",
  unexpressed = 0)
}
\arguments{
\item{object}{RangedSummarizedExperiment: CAGE data quantified at CTSS, cluster or gene-level.}

\item{inputAssay}{character: Name of assay holding input expression values.}

\item{outputColumn}{character: Name of column in rowRanges to hold support values.}

\item{unexpressed}{numeric: Support will be calculated based on features larger than this cutoff.}
}
\value{
RangedSummarizedExperiment with support added as a column in rowRanges.
}
\description{
Calculate support of CAGE data.
}
