% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSE_function.R
\name{calcTPM}
\alias{calcTPM}
\title{Calculate TPM for a CTSS file}
\usage{
calcTPM(gr)
}
\arguments{
\item{gr}{GRanges: Single CTSS file.}
}
\value{
GRanges with column "tpm" added.
}
\description{
Adds a column "tpm" to a CTSS files containing (naive) Tag-per-million score for each bp position
}
\examples{
# ADD_EXAMPLES_HERE
}

