% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bidirectional_clustering.R
\name{calculateBidirectionality}
\alias{calculateBidirectionality}
\title{Calculate sample-wise bidirectional transcription}
\usage{
calculateBidirectionality(ctss, features, biocParallel = bpparam())
}
\arguments{
\item{ctss}{GRangesList or SimpleList: GRanges with CTSSs in the score column.}

\item{features}{GRanges: Bidirectional sites (with the maximal bidirectional site in the thick column).}

\item{biocParallel}{BiocParallelParam: Settings for parallel backend.}
}
\value{
Integer-vector of same length as features with the number of samples showing bidirectional transcription.
}
\description{
For each bidirectional site, calculates how many individual samples shows transcription in both directions.
}
\examples{
# ADD_EXAMPLES_HERE
}
