% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{calculateShape}
\alias{calculateShape}
\title{Calculate TC shape}
\usage{
calculateShape(ctssCoverage, TCs, shapeFunction, ...)
}
\arguments{
\item{ctssCoverage}{GRanges: GRanges with CTSS as score column.}

\item{TCs}{GRanges: TCs to be analyzed.}

\item{shapeFunction}{function: Function to apply to each TC (See details).}

\item{...}{additional arguments passed to shapeFunction.}
}
\value{
Numeric vector of same length as TCs holding output from shapeFunction.
}
\description{
Apply a shape-function to the global coverage of every TC.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other Shape functions: \code{\link{InterQuartileRange}},
  \code{\link{ShannonEntropy}}, \code{\link{sumRounding}}
}
