% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwIO.R
\name{checkBigWigGenomes}
\alias{checkBigWigGenomes}
\title{Check if BigWig-files are compatible with a genome.}
\usage{
checkBigWigGenomes(bwPlus, bwMinus, seqInfo)
}
\arguments{
\item{bwPlus}{BigWigFileList: BigWig files with plus-strand CTSS data.}

\item{bwMinus}{BigWigFileList: BigWig files with minus-strand CTSS data.}

\item{seqInfo}{Seqinfo: Genome information.}
}
\value{
None: Throws an error if the supplied genome is incompabtible.
}
\description{
Given a genome, checks whether a series of BigWig-files are compatible by checking if all common seqlevels have equal seqlengths.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other IO functions: \code{\link{commonBigWigGenome}},
  \code{\link{fastReadBigWig}}, \code{\link{readCTSS}},
  \code{\link{uniqueBigWigGenomes}}
}
