% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkPeaked}
\alias{checkPeaked}
\title{Helper for checking cluster with peaks}
\usage{
checkPeaked(object)
}
\arguments{
\item{object}{GRanges or GPos: Clusters with peaks to be checked.}
}
\value{
TRUE if object is correct format, otherwise an error is thrown
}
\description{
Checks whether a supplied set of cluster have valid peaks: Whether the thick
column contains IRanges all contained within the main ranges.
}
\examples{
data(exampleUnidirectional)
checkPeaked(rowRanges(exampleUnidirectional))
}
\seealso{
Other Checking functions: 
\code{\link{checkPooled}()}
}
\concept{Checking functions}
