% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkPooled}
\alias{checkPooled}
\title{Helper for checking pooled signal}
\usage{
checkPooled(object)
}
\arguments{
\item{object}{GRanges or GPos: Pooled signal to be checked}
}
\value{
TRUE if object is correct format, otherwise an error is thrown
}
\description{
Checks whether a supplied pooled signal is valid: Single bp disjoint with
signal in the score column with supplied genome information.
}
\examples{
data(exampleCTSSs)
checkPooled(rowRanges(exampleCTSSs))
}
\seealso{
Other Checking functions: 
\code{\link{checkPeaked}()}
}
\concept{Checking functions}
