% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwIO.R
\name{commonBigWigGenome}
\alias{commonBigWigGenome}
\title{Find a common genome for a series of BigWig files.}
\usage{
commonBigWigGenome(bwPlus, bwMinus, method = "intersect")
}
\arguments{
\item{bwPlus}{BigWigFileList: BigWig files with plus-strand CTSS data.}

\item{bwMinus}{BigWigFileList: BigWig files with minus-strand CTSS data.}

\item{method}{character: Either "intersect" or "union".}
}
\value{
Seqinfo-object.
}
\description{
Finds a common genome for a series of BigWig-files, either using only levels present in all files (intersect) or in any file (union).
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other IO functions: \code{\link{checkBigWigGenomes}},
  \code{\link{fastReadBigWig}}, \code{\link{readCTSS}},
  \code{\link{uniqueBigWigGenomes}}
}
