% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantification.R
\name{countScoredOverlaps}
\alias{countScoredOverlaps}
\title{Internal function: countOverlaps with scores}
\usage{
countScoredOverlaps(gr, features)
}
\arguments{
\item{gr}{GRanges: Ranges with scores.}

\item{features}{GRanges: Ranges to be quantified.}
}
\value{
numeric vector of same length as features.
}
\description{
Similar to countOverlaps, but takes into account the scores column.
}
\examples{
# ADD_EXAMPLES_HERE
}

