% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverageOfCTSS}
\alias{coverageOfCTSS}
\alias{coverageOfCTSS.GRangesList}
\alias{coverageOfCTSS.character}
\alias{coverageOfCTSS.default}
\title{Calculate CTSS coverage}
\usage{
coverageOfCTSS(ctss, ctssCutoff = 0)

\method{coverageOfCTSS}{default}(ctss, ctssCutoff = 0)

\method{coverageOfCTSS}{GRangesList}(ctss, ctssCutoff = 0)

\method{coverageOfCTSS}{character}(ctss, ctssCutoff = 0)
}
\arguments{
\item{ctss}{GRangesList or character: CTSS data stores as a GRangesList or a vector of paths to CTSS-files.}

\item{ctssCutoff}{integer: Trim away positions with this or less CAGE-tags.}
}
\value{
GRanges with bp-wise TPM-coverage.
}
\description{
Calculates TPM-coverage for a series of CTSS-files.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default method (always throws an error)

\item \code{GRangesList}: High memory method

\item \code{character}: Low memory method
}}
\examples{
# ADD_EXAMPLES_HERE
}

