% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{extendRanges}
\alias{extendRanges}
\title{Extend GRanges upstrean and/or downstream}
\usage{
extendRanges(gr, upstream = 0, downstream = 0)
}
\arguments{
\item{gr}{GRanges: Ranges to be extended, any seqinfo will be used to trim out-of-bounds ranges.}

\item{upstream}{integer: Number of basepairs to extend upstream.}

\item{downstream}{integer: Number of basepairs to extend downstream.}
}
\value{
GRanges with ranges extended.
}
\description{
Extends the ranges of GRanges upstream and/or downstream, while retaining associated information in mcols.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other Helper functions: \code{\link{TPM}},
  \code{\link{resolvedOverlaps}},
  \code{\link{splitByStrand}}, \code{\link{swapRanges}}
}
\author{
Hervé Pagès
}
