% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{fastReadBED}
\alias{fastReadBED}
\title{Internal function: Quickly read simple BED-file}
\usage{
fastReadBED(fname)
}
\arguments{
\item{fname}{character: Path to BED file.}
}
\value{
data.table
}
\description{
Uses fread from data.table to quickly reading a simple bed file. Only scores are kept, any names are discarded
}
\examples{
# ADD_EXAMPLES_HERE
}

