% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwIO.R
\name{fastReadBigWig}
\alias{fastReadBigWig}
\title{Import a pair of BigWig-files.}
\usage{
fastReadBigWig(fnamePlus, fnameMinus, seqInfo = NULL)
}
\arguments{
\item{fnamePlus}{BigWigFile: BigWig file with plus-strand CTSS data.}

\item{fnameMinus}{BigWigFile: BigWig file with minus-strand CTSS data.}

\item{seqInfo}{Seqinfo or NULL: If not NULL, use the supplied Seqinfo as genome.}
}
\value{
GRanges with CTSS info as the score column.
}
\description{
Reads in CTSS-data stored in a BigWig-file for each strand.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other IO functions: \code{\link{checkBigWigGenomes}},
  \code{\link{commonBigWigGenome}}, \code{\link{readCTSS}},
  \code{\link{uniqueBigWigGenomes}}
}

