% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSE_function.R
\name{findTagClusters}
\alias{findTagClusters}
\title{Simple Tag-Clustering}
\usage{
findTagClusters(ctss, tpmCutoff = 1, mergeDist = 25)
}
\arguments{
\item{ctss}{GRangesList: CTSS-GRanges.}

\item{tpmCutoff}{numeric: BP-positions with this or less TPM will not be used for tag clustering.}

\item{mergeDist}{integer: TCs within this distance of eachother are merged.}
}
\value{
GRanges of TCs, with TC peak position and summed TPM expression.
}
\description{
Generates simple Tag Clusters (TCs) from CTSS-GRanges: Calculates TPM-coverage across all samples and finds cluster above a certain cutoff, merging cluster within a specified distance.
}
\examples{
# ADD_EXAMPLES_HERE
}

