% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{HartigansDip}
\alias{HartigansDip}
\title{Hartigans' dip statistics}
\usage{
HartigansDip(x, roundingFun = runningRounding)
}
\arguments{
\item{x}{numeric Rle vector: Coverage series.}

\item{roundingFun}{function: Function for rounding an Rle vector.}
}
\value{
numeric.
}
\description{
Calculates the Hartigans Dip Statistics for a vector, by first rounding values to integers. A dip statistics of 0 indicates a unimodal distribtuion and a dip statistics of 1 indicates a bi- or multimodal distribution.
}
\examples{
# ADD_EXAMPLES_HERE
}
