% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{InterQuartileRange}
\alias{InterQuartileRange}
\title{Calculate InterQuartile Range}
\usage{
InterQuartileRange(x, lower = 0.25, upper = 0.75)
}
\arguments{
\item{x}{Numeric: values.}

\item{lower}{numeric: Lower quartile.}

\item{upper}{numeric: Upper quartile.}
}
\value{
Numeric
}
\description{
Calculates the interquartile range of a vector.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other Shape functions: \code{\link{ShannonEntropy}},
  \code{\link{calculateShape}}
}

