% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantification.R
\name{lowMemOverlap}
\alias{lowMemOverlap}
\title{Internal function: Low memory CTSS overlap}
\usage{
lowMemOverlap(fname, features)
}
\arguments{
\item{fname}{character: CTSS-BED file}

\item{features}{GRanges: Features to be quantified.}
}
\value{
numeric vector with the same length as gr
}
\description{
Quickly read in a CTSS file prior to overlap with countScoredOverlaps
}
\examples{
# ADD_EXAMPLES_HERE
}

