% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{mergeCoverage}
\alias{mergeCoverage}
\title{Internal function: Merge data.tables with coverage}
\usage{
mergeCoverage(dt1, dt2)
}
\arguments{
\item{dt1}{data.table: Coverage in a data.table}

\item{dt2}{data.table: Coverage in a data.table}
}
\value{
Coverage in a data.table
}
\description{
To be used in a Reduced-loop, merge data.tables containing coverage information.
}
\examples{
# ADD_EXAMPLES_HERE
}

