% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSE_function.R
\name{overlapCTSS}
\alias{overlapCTSS}
\title{Count overlaps for a CTSS file}
\usage{
overlapCTSS(tcs, ctss)
}
\arguments{
\item{tcs}{GRanges: Genomic Ranges to count overlaps}

\item{ctss}{GRanges: CTSS-GRanges}
}
\value{
integer vector of same length as tcs: Number of tags in each genomic range.
}
\description{
Counts overlaps of a CTSS-GRanges with a set of GRanges.
}
\examples{
# ADD_EXAMPLES_HERE
}

