% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSE_function.R
\name{quantifyFeatures}
\alias{quantifyFeatures}
\title{Quantify CTSS file into an Expression Matrix (EM)}
\usage{
quantifyFeatures(tcs, ctss)
}
\arguments{
\item{tcs}{GRanges: TC coordinates.}

\item{ctss}{GRangesList: Set of CTSS-GRanges.}
}
\value{
Integer matrix of dimensions: Number of TCs x Number of samples.
}
\description{
For a set of CTSS-GRanges and a set of Tag Cluster (TCs), count number of CAGE-tags in each TC in each sample and return an expression matrix.
}
\examples{
# ADD_EXAMPLES_HERE
}

