% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.R
\name{quickEnhancers}
\alias{quickEnhancers}
\title{Identify and quantify enhancers.}
\usage{
quickEnhancers(object)
}
\arguments{
\item{object}{RangedSummarizedExperiment: Location and counts of CTSSs, usually found by calling quantifyCTSSs.}
}
\value{
RangedSummarizedExperiment containing location and counts of enhancers.
}
\description{
A convienient wrapper around clusterBidirectionally, subsetByBidirectionality and quantifyClusters.
}
