% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.R
\name{quickGenes}
\alias{quickGenes}
\title{Identify and quantify genes.}
\usage{
quickGenes(object, geneModels = NULL, ...)
}
\arguments{
\item{object}{RangedSummarizedExperiment: Location and counts of clusters,
usually found by calling quantifyClusters.}

\item{geneModels}{TxDb or GRanges: Gene models via a TxDb, or manually
specified as a GRangesList.}

\item{...}{additional arguments passed to assignGeneID.}
}
\value{
RangedSummarizedExperiment containing gene expression and clusters
  assigned within each gene.
}
\description{
A convienient wrapper around assignGeneID, and quantifyGenes. Also removes
unstranded features
}
\examples{
# See the CAGEfightR vignette for an overview!
}
\seealso{
Other Wrapper functions: 
\code{\link{quickEnhancers}()},
\code{\link{quickTSSs}()}
}
\concept{Wrapper functions}
