% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.R
\name{quickTSSs}
\alias{quickTSSs}
\title{Identify and quantify Transcription Start Sites (TSSs).}
\usage{
quickTSSs(object)
}
\arguments{
\item{object}{RangedSummarizedExperiment: Location and counts of CTSSs,
usually found by calling quantifyCTSSs.}
}
\value{
RangedSummarizedExperiment containing location and counts of TSSs
}
\description{
A convienient wrapper around calcTPM, calcPooled, tuneTagClustering,
clusterUnidirectionally and quantifyClusters.
}
\examples{
# See the CAGEfightR vignette for an overview!
}
\seealso{
Other Wrapper functions: 
\code{\link{quickEnhancers}()}
}
\concept{Wrapper functions}
