% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSE_function.R
\name{readCTSS}
\alias{readCTSS}
\title{Import CTSS file}
\usage{
readCTSS(fname)
}
\arguments{
\item{fname}{character: Path to ctss file on disk.}
}
\value{
GRanges representation of CTSS file
}
\description{
Uses the `fread`-function from the data.table package to quickly import a CTSS file from disk into a GRanges object.
}
\examples{
# ADD_EXAMPLES_HERE
}

