% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSE_function.R
\name{readMultipleCTSS}
\alias{readMultipleCTSS}
\title{Read multiple CTSS files}
\usage{
readMultipleCTSS(fnames)
}
\arguments{
\item{fnames}{character: Vector of file names.}
}
\value{
GRangesList
}
\description{
Uses the `fread`-function from the data.table package to quickly import multiple CTSS files from disk into a GRangesList object.
}
\examples{
# ADD_EXAMPLES_HERE
}

