% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{reduceCoverage}
\alias{reduceCoverage}
\title{Internal function: Reduce-style coverage calculation}
\usage{
reduceCoverage(ctss, ctssCutoff = 0, progressBar = FALSE)
}
\arguments{
\item{ctss}{character: CTSS-BED files}

\item{ctssCutoff}{integer: Trim away positions with this or less CAGE-tags.}

\item{progressBar}{logical: Whether to track progress.}
}
\value{
data.table
}
\description{
Uses a reduce approach to calculate coverage. Each file is TPM-transformed prior to being reduced.
}
\examples{
# ADD_EXAMPLES_HERE
}

