% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{relabelUnannotated}
\alias{relabelUnannotated}
\title{Relabel unannotated genes}
\usage{
relabelUnannotated(geneIDs, prefix = "Novel")
}
\arguments{
\item{geneIDs}{character: Gene names, with NA indicating unannotated.}

\item{prefix}{character: Prefix to add before counter.}
}
\value{
character vector of the same length as geneID
}
\description{
Relabels unannotated genes, i.e. as "Novel1". Usesful when testing for Alternative Transcript Usage (ATU).
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other Annotation functions: \code{\link{assignGeneID}},
  \code{\link{assignTxID}}, \code{\link{assignTxType}},
  \code{\link{simplifyTxTypes}}, \code{\link{sumOverGenes}}
}
