% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{resolvedOverlaps}
\alias{resolvedOverlaps}
\title{Finding overlapping genomic ranges resolved for multiple overlaps.}
\usage{
resolvedOverlaps(query, subject, select, ...)
}
\arguments{
\item{query}{GRanges or GRangesList: Passed directly to findOverlaps.}

\item{subject}{GRanges or GRangesList: Passed directly to findOverlaps.}

\item{select}{character: How to resolve multiple overlaps: findOverlap's "first", "last", "arbitrary" or additional "shortest" or "longest".}

\item{...}{additional arguments passed to findOverlaps.}
}
\value{
Similar to findOverlaps with select="first", "last" or "arbitrary".
}
\description{
Similar to findOverlaps, but provides to additional methods for resolving the case of a query overlapping multiple subjects using subject lenghts.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other Helper functions: \code{\link{TPM}},
  \code{\link{extendRanges}}, \code{\link{splitByStrand}},
  \code{\link{swapRanges}}
}
