% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{runningRounding}
\alias{runningRounding}
\title{Round Rle vector to integers while preserving running sum}
\usage{
runningRounding(x)
}
\arguments{
\item{x}{numeric Rle vector: Coverage series.}
}
\value{
Integer vector.
}
\description{
Round an Rle vector, while preserving the running sum. Slightly faster than sumRounding. Can be used with the HartigansDip function.
}
\examples{
# ADD_EXAMPLES_HERE
}
\references{
\url{https://stackoverflow.com/a/36970695}
}
