% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSE_function.R
\name{scoreAsTPM}
\alias{scoreAsTPM}
\title{Rescales CTSS score from counts to TPM}
\usage{
scoreAsTPM(gr)
}
\arguments{
\item{gr}{GRanges: Single CTSS file.}
}
\value{
GRanges with scores in TPM
}
\description{
Rescales the scores of a CTSS-GRanges to Tag-per-million (TPM)
}
\examples{
# ADD_EXAMPLES_HERE
}

