% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{ShannonEntropy}
\alias{ShannonEntropy}
\title{Calculate Shannon Entropy}
\usage{
ShannonEntropy(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
Numeric.
}
\description{
Calculates the Shannon Entropy (base log2) for a vector. Zeros are removed before calculation.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other Shape functions: \code{\link{InterQuartileRange}},
  \code{\link{calculateShape}}
}

