% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{shapeEntropy}
\alias{shapeEntropy}
\title{Shape statistic: Shannon Entropy}
\usage{
shapeEntropy(x)
}
\arguments{
\item{x}{numeric Rle vector: Coverage series.}
}
\value{
Numeric.
}
\description{
Calculates the Shannon Entropy (base log2) for a vector. Zeros are removed before calculation.
}
\seealso{
Other Shape functions: \code{\link{calcShape}},
  \code{\link{shapeIQR}}, \code{\link{shapeMultimodality}}
}
