% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{shapeIQR}
\alias{shapeIQR}
\title{Shape statitic: Interquartile range}
\usage{
shapeIQR(x, lower = 0.25, upper = 0.75)
}
\arguments{
\item{x}{numeric Rle vector: Coverage series.}

\item{lower}{numeric: Lower quartile.}

\item{upper}{numeric: Upper quartile.}
}
\value{
Numeric
}
\description{
Calculates the interquartile range of a vector.
}
\seealso{
Other Shape functions: \code{\link{calcShape}},
  \code{\link{shapeEntropy}},
  \code{\link{shapeMultimodality}}
}
