% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{shapeMean}
\alias{shapeMean}
\title{Shape statistic: Mean}
\usage{
shapeMean(x)
}
\arguments{
\item{x}{numeric Rle vector: Coverage series.}
}
\value{
Numeric
}
\description{
Calculates the mean of a vector.
}
\examples{
# Hypothetical shard/broad clusters:
x_sharp <- Rle(c(1,1,1,4,5,2,1,1))
x_broad <- Rle(c(1,2,3,5,4,3,2,1))

# Calculate mean
shapeMean(x_sharp)
shapeMean(x_broad)

# See calcShape for more usage examples
}
\seealso{
Other Shape functions: 
\code{\link{calcShape}()},
\code{\link{shapeEntropy}()},
\code{\link{shapeIQR}()}
}
\concept{Shape functions}
