% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{shapeMultimodality}
\alias{shapeMultimodality}
\title{Shape statistic: Multimodality}
\usage{
shapeMultimodality(x)
}
\arguments{
\item{x}{numeric Rle vector: Coverage series.}
}
\value{
Numeric.
}
\description{
Shape statistic: Multimodality
}
\examples{
# See calcShape for usage examples
}
\seealso{
Other Shape functions: 
\code{\link{calcShape}()},
\code{\link{shapeEntropy}()},
\code{\link{shapeIQR}()}
}
\concept{Shape functions}
