% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{simpleRounding}
\alias{simpleRounding}
\title{Round Rle vector to integers}
\usage{
simpleRounding(x)
}
\arguments{
\item{x}{numeric Rle vector.}
}
\value{
Integer vector.
}
\description{
Wrapper for round(). Can be used with the HartigansDip function.
}
\examples{
# ADD_EXAMPLES_HERE
}
