% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_clustering.R
\name{simpleTagClustering}
\alias{simpleTagClustering}
\title{Simple Tag Clustering - DEPRECATED}
\usage{
simpleTagClustering(ctssCoverage, tpmCutoff = 0, mergeDist = 20)
}
\arguments{
\item{ctssCoverage}{GRanges: CTSS coverage.}

\item{tpmCutoff}{numeric: Minium TPM coverage to be considered as TC.}

\item{mergeDist}{integer: Merge TCs within this distance.}
}
\value{
GRanges with TPM sum as the score column, and TC peak as the thick column.
}
\description{
Finds Tag Cluster (TCs) with a TPM above a certain threshold. Addtionally calculates the sum and peak position of the TCs.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other Tag-clustering functions: \code{\link{tagClustering}},
  \code{\link{tuneTagClustering}}
}
