% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{simplifyTxTypes}
\alias{simplifyTxTypes}
\title{Simplify transcript types}
\usage{
simplifyTxTypes(txTypes, scheme = "genic")
}
\arguments{
\item{txTypes}{factor: txTypes returned by asignTxTypes-function.}

\item{scheme}{character or list: Scheme for simplification: Currently supports "genic" and "promoter proximity. Advanced users may directly supply a named list of levels.}
}
\value{
factor with simplified txTypes
}
\description{
Simplify annotation of transcript types (txTypes).
}
\examples{
# ADD_EXAMPLES_HERE
}
