% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splitByStrand}
\alias{splitByStrand}
\title{Split by strand.}
\usage{
splitByStrand(object)
}
\arguments{
\item{object}{object with strand information (GRanges, Gpos, RangedSummarizedExperiment, etc.).}
}
\value{
List with three elements: +, - and *.
}
\description{
Splits an object into a list by strand
}
