% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSE.R
\name{stackRSEs}
\alias{stackRSEs}
\title{Stack two RangedSummarizedExperiments}
\usage{
stackRSEs(a, b, removeIfOverlapping = "both")
}
\arguments{
\item{a}{RangedSummarizedExperiment: RSE to be stacked.}

\item{b}{RangedSummarizedExperiment: RSE to be stacked.}

\item{removeIfOverlapping}{character: Whether to discard features in either "a" or "b" if they overlap. "both" does not remove overlapping features.}
}
\value{
Stacked and sorted RangedSummarizedExperiment.
}
\description{
A convienient wrapper around rbind for stacking rows RangedSummarizedExperiment-objecs (RSEs). Automatically keeps only shared rowData and assays, and has the option of discarding overlapping features. Useful for combing RSEs containing TSSs and enhancers.
}
\examples{
# ADD_EXAMPLES_HERE
}
