% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_browser.R
\name{strandedCoverageTrack}
\alias{strandedCoverageTrack}
\title{Stranded Coverage Genome Browser Track}
\usage{
strandedCoverageTrack(gr, plusColor = "cornflowerblue",
  minusColor = "tomato", ...)
}
\arguments{
\item{gr}{GRanges: GRanges with coverage in the score column.}

\item{plusColor}{character: Color for plus-strand coverage.}

\item{minusColor}{character: Color for minus-strand coverage.}

\item{...}{additional arguments passed on to DataTrack.}
}
\value{
DataTrack-object.
}
\description{
Create a Gviz-track with stranded coverage (Plus/minus coverage as positive/negative). This representation makes it easy to identify bidirectional peaks.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other Genome Browser functions: \code{\link{tagClusterTrack}}
}
