% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{sumByGene}
\alias{sumByGene}
\title{Collapse an Expression Matrix by summing over genes}
\usage{
sumByGene(em, genes, keepUnannotated = FALSE, prefix = "Novel")
}
\arguments{
\item{em}{matrix: Expression matrix.}

\item{genes}{factor or character: Vector of gene names to group by.}

\item{keepUnannotated}{logical: Whether to discard (FALSE) or keep (TRUE) un annotated transcripts.}

\item{prefix}{character: If keepUnannotated is TRUE, this prefix is added to all unannotated transcripts along with a counter.}
}
\value{
Expression matrix summed by genes.
}
\description{
Summarise a transcript-level EM to a gene-level EM, i.e. for GO-term analysis.
}
\examples{
# ADD_EXAMPLES_HERE
}

