% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSE_function.R
\name{summarizeCAGE}
\alias{summarizeCAGE}
\title{Summarize CAGE experiment.}
\usage{
summarizeCAGE(ctss, design, fun = simpleTCs, ...)
}
\arguments{
\item{ctss}{GRangesList: CTSS-GRanges.}

\item{design}{data.frame: Study design.}

\item{fun}{function: Function for finding TCs (See Details for additional information on customizing.).}

\item{...}{additional arguments passed to fun}
}
\value{
RangedSummarizedExperiment.
}
\description{
This function is a user-friendly wrapper multiple other functions. Given a set of CTSS-GRanges, will find Tag Cluster (TCs) and quantify expression within them. Results are returned as a RangedSummarizedExperiment (RSE).
}
\details{
TBA: Description of custom clustering
}
\note{
This function does not sort or match the input, so arguments must be provided with matching columns, rows, etc.
}
\examples{
# ADD_EXAMPLES_HERE
}

