% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{sumOverGenes}
\alias{sumOverGenes}
\title{Collapse Expression Matrix and Ranges by summing over genes}
\usage{
sumOverGenes(RSE, inputAssay, geneID, calcStats = TRUE)
}
\arguments{
\item{RSE}{RangedSummarizedExperiment: Transcript-level data to be summarized.}

\item{inputAssay}{character: Name of assay holding expression values to be summed, i.e. "counts".}

\item{geneID}{character: Name of column in rowData holding geneIDs.}

\item{calcStats}{logical: Whether to calculate extra statistics for the merged transcripts (i.e. number of transcripts in each gene).}
}
\value{
RangedSummarizedExperiment with inputAssay summed over genes and rowRanges as a GRangesList. If calcStats=TRUE, rowData contains gene-wise statistics. colData and metaData are unchanged from the original RangedSummarizedExperiment.
}
\description{
Summarize a transcript-level EM to a gene-level EM,  i.e. for GO-term analysis. Transcripts are grouped within genes. Unannotated transcripts (NAs) are discarded.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other Annotation functions: \code{\link{assignGeneID}},
  \code{\link{assignTxID}}, \code{\link{assignTxType}},
  \code{\link{relabelUnannotated}},
  \code{\link{simplifyTxTypes}}
}
