% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{sumRounding}
\alias{sumRounding}
\title{Round Rle vector to integers while preserving sum}
\usage{
sumRounding(x)
}
\arguments{
\item{x}{numeric Rle vector: Coverage series.}
}
\value{
Integer vector.
}
\description{
Round an Rle vector, while preserving the sum of the vector as closely as possible. Slightly slower than runningRounding. Can be used with the HartigansDip function.
}
\examples{
# ADD_EXAMPLES_HERE
}
\references{
\url{https://stackoverflow.com/a/32544987}
}
\seealso{
Other Shape functions: \code{\link{InterQuartileRange}},
  \code{\link{ShannonEntropy}},
  \code{\link{calculateShape}}
}
