% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwCoverage.R
\name{supportOfCTSS}
\alias{supportOfCTSS}
\title{Calculate global CTSS support.}
\usage{
supportOfCTSS(grl)
}
\arguments{
\item{grl}{GRangesList or SimpleList: GRanges with CTSSs in the score column.}
}
\value{
GRanges with global support as the score-column.
}
\description{
Given a series of GRanges calculates the support of each basepair as the number of samples with non-zero counts.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other Coverage functions: \code{\link{coverageOfCTSS}},
  \code{\link{trimAndNormalizeGR}}
}
