% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwCoverage.R
\name{TPM}
\alias{TPM}
\title{Calculate TPM.}
\usage{
TPM(x)
}
\arguments{
\item{x}{Integer: Counts to be TPM normalized.}
}
\value{
Numeric vector of same length as x.
}
\description{
Calculate Tags-Per-Million (TPM) for a numeric vector.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other Helper functions: \code{\link{splitByStrand}}
}

