% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\docType{methods}
\name{trackBalance}
\alias{trackBalance}
\alias{trackBalance,GenomicRanges-method}
\alias{trackBalance,RangedSummarizedExperiment-method}
\title{Create Genome Browser Track of bidirectional balance scores}
\usage{
trackBalance(object, ...)

\S4method{trackBalance}{GenomicRanges}(object, window = 199,
  plusColor = "cornflowerblue", minusColor = "tomato",
  balanceColor = "forestgreen", ...)

\S4method{trackBalance}{RangedSummarizedExperiment}(object, ...)
}
\arguments{
\item{object}{GenomicRanges or RangedSummarizedExperiment: Ranges with CTSSs in the score column.}

\item{...}{additional arguments passed to DataTrack.}

\item{window}{integer: Width of sliding window used for calculating windowed sums.}

\item{plusColor}{character: Color for plus-strand coverage.}

\item{minusColor}{character: Color for minus-strand coverage.}

\item{balanceColor}{character: Color for bidirectional balance.}
}
\value{
list of 3 DataTracks for upstream, downstream and balance.
}
\description{
Visualize balance scores used for detectiong of bidirectional sites. Mainly intended as diagnostic tools for expert user.
}
\note{
Potentially consumes a large amount of memory!
}
