% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\docType{methods}
\name{trackCTSS}
\alias{trackCTSS}
\alias{trackCTSS,GenomicRanges-method}
\alias{trackCTSS,RangedSummarizedExperiment-method}
\alias{trackCTSS,GPos-method}
\title{Create Genome Browser track of CTSSs.}
\usage{
trackCTSS(object, ...)

\S4method{trackCTSS}{GenomicRanges}(object, plusColor = "cornflowerblue",
  minusColor = "tomato", ...)

\S4method{trackCTSS}{RangedSummarizedExperiment}(object, ...)

\S4method{trackCTSS}{GPos}(object, ...)
}
\arguments{
\item{object}{GenomicRanges or RangedSummarizedExperiment: Ranges with CTSSs in the score column.}

\item{...}{additional arguments passed on to DataTrack.}

\item{plusColor}{character: Color for plus-strand coverage.}

\item{minusColor}{character: Color for minus-strand coverage.}
}
\value{
DataTrack-object.
}
\description{
Create a Gviz-track of CTSSs, where Plus/minus strand signal is shown positive/negative. This representation makes it easy to identify bidirectional peaks.
}
\seealso{
Other Genome Browser functions: \code{\link{trackClusters}}
}
