% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwCoverage.R
\name{trimAndNormalizeGR}
\alias{trimAndNormalizeGR}
\title{Basic preprocessing of CTSS-files.}
\usage{
trimAndNormalizeGR(gr, preFilterPositions = NULL, preFilterCTSS = NULL,
  preFilterTPM = NULL)
}
\arguments{
\item{gr}{GRanges: GRanges holding CTSS counts in the score column.}

\item{preFilterPositions}{GRanges: Ranges to discard from calculations.}

\item{preFilterCTSS}{integer: Minimum count for retained CTSSs.}

\item{preFilterTPM}{numeric: Minimum TPM-value for retained CTSSs.}
}
\value{
GRanges with TPM-values in the score column.
}
\description{
Function that trims away weakly expressed CTSSs by discarding based on CTSS counts and/or TPM-values, before performing a final TPM-calculation.
}
\note{
The order of trimming is: Optional trimming on CTSS counts, optional trimming of TPM-values, final calculation of TPM.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other Coverage functions: \code{\link{coverageOfCTSS}},
  \code{\link{supportOfCTSS}}
}
