% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_trimming.R
\name{trimAroundPeaks}
\alias{trimAroundPeaks}
\title{Trim width of TCs by distance from TC peak}
\usage{
trimAroundPeaks(TCs, ctssCoverage, upstream, downstream, peaks = "thick")
}
\arguments{
\item{TCs}{GRanges: TCs to be trimmed.}

\item{ctssCoverage}{GRanges: CTSS coverage.}

\item{upstream}{integer: Maximum upstream distance from TC peak.}

\item{downstream}{integer: Maximum downstream distance from TC peak.}

\item{peaks}{character: Name of column in TCs holding TC peaks as an IRanges.}
}
\value{
GRanges with trimmed TCs, including recalculated peaks and scores.
}
\description{
Trim the width of TCs by distance from the TC peaks.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other TC trimming functions: \code{\link{trimToPercentiles}}
}
