% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_trimming.R
\name{trimToPercentiles}
\alias{trimToPercentiles}
\title{Trim width of TCs to expression percentiles}
\usage{
trimToPercentiles(TCs, ctssCoverage, percentile = 0.1, symmetric = FALSE)
}
\arguments{
\item{TCs}{GRanges: TCs to be trimmed.}

\item{ctssCoverage}{GRanges: CTSS coverage.}

\item{percentile}{numeric: Fraction of expression to remove from TCs.}

\item{symmetric}{logical: Whether to trim the same amount from both edges of the TC (TRUE) or always trim from the least expressed end (FALSE).}
}
\value{
GRanges with trimmed TCs, including recalculated peaks and scores.
}
\description{
Given a set of TCs and genome-wide CTSS coverage, reduce the width of TC until a certain amount of expression has been removed.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other TC trimming functions: \code{\link{trimAroundPeaks}}
}
