% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_clustering.R
\name{tuneTagClustering}
\alias{tuneTagClustering}
\title{Determine optimal expression threshold for Tag Clustering.}
\usage{
tuneTagClustering(ctssCoverage, stepSize = 1, mergeDist = 20,
  biocParallel = bpparam())
}
\arguments{
\item{ctssCoverage}{GRanges: CTSS coverage in score column.}

\item{stepSize}{integer: Size of steps for increasing threshold}

\item{mergeDist}{integer: Merge TCs within this distance.}

\item{biocParallel}{BiocParallelParam: Settings for parallel backend.}
}
\value{
data.frame with two columns: threshold and nTCs (number of Tag Clusters)
}
\description{
This function counts the number of Tag Clusters (TCs) for an exponentially increasing series of expression cutoffs.
}
\examples{
# ADD_EXAMPLES_HERE
}
