% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwIO.R
\name{uniqueBigWigGenomes}
\alias{uniqueBigWigGenomes}
\title{Count the number of different genomes in a series of BigWig-files.}
\usage{
uniqueBigWigGenomes(bwPlus, bwMinus)
}
\arguments{
\item{bwPlus}{BigWigFileList: BigWig files with plus-strand CTSS data}

\item{bwMinus}{BigWigFileList: BigWig files with minus-strand CTSS data}
}
\value{
SimpleList holding Seqinfo-objects.
}
\description{
Find all unique genomes in in a series of BigWig-files.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other IO functions: \code{\link{checkBigWigGenomes}},
  \code{\link{commonBigWigGenome}},
  \code{\link{fastReadBigWig}}, \code{\link{readCTSS}}
}
