% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{assignGeneID}
\alias{assignGeneID}
\title{Annotate GRanges with gene ID}
\usage{
assignGeneID(gr, txdb, proximalUpstream = 1000, proximalDownstream = 100,
  multiHits = "arbitrary")
}
\arguments{
\item{gr}{GRanges: Ranges to annotated.}

\item{txdb}{TxDb: Transcript database to use for annotation}

\item{proximalUpstream}{integer: Maximum distance upstream of annotated promoter to be considered part of genes.}

\item{proximalDownstream}{integer: Maximum distance downstream of annotated promoter to be considered part of genes.}

\item{multiHits}{character: How to choose when multiple genes overlap.}
}
\value{
character vector of same length as gr containing Entrez gene IDs for each range.
}
\description{
Annotate a set of ranges in a GRanges object with Entrez Gene Identifier (ID) based on their genic context. Genes are obtained from a TxDb object.
}
\examples{
# ADD_EXAMPLES_HERE
}

