% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{assignGeneID}
\alias{assignGeneID}
\title{Annotate GRanges with gene ID}
\usage{
assignGeneID(gr, txdb, upstream = 1000, downstream = 0,
  resolveRule = "shortest")
}
\arguments{
\item{gr}{GRanges: Ranges to annotated.}

\item{txdb}{TxDb: Transcript database to use for annotation.}

\item{upstream}{integer: Region upstream of gene to be considered part of the gene.}

\item{downstream}{integer: Region downstream of gene to be considered part of the gene.}

\item{resolveRule}{character: How to resolve multiple overlaps: "first", "last", "arbitrary", "shortest" or "longest".}
}
\value{
character vector of same length as gr containing a single gene ID for each range.
}
\description{
Annotate a set of ranges in a GRanges object with gene IDs (i.e. Entrez Gene Identifiers) based on their genic context. Genes are obtained from a TxDb object.
}
\examples{
# ADD_EXAMPLES_HERE
}
