% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{assignTxID}
\alias{assignTxID}
\title{Annotate GRanges with transcript ID}
\usage{
assignTxID(gr, txdb, upstream = 1000, downstream = 100)
}
\arguments{
\item{gr}{GRanges: Ranges to annotated.}

\item{txdb}{TxDb: Transcript database to use for annotation}

\item{upstream}{integer: Maximum distance upstream of annotated TSS to be considered part of genes.}

\item{downstream}{integer: Maximum distance downstream of annotated TSS to be considered part of genes.}
}
\value{
character vector of same length as gr containing transcript names.
}
\description{
Annotate a set of ranges in a GRanges object with Transcript Identifier (TxID) based on their genic context. Transcripts are obtained from a TxDb object.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other Annotation functions: \code{\link{assignGeneID}},
  \code{\link{assignTxType}},
  \code{\link{relabelUnannotated}},
  \code{\link{simplifyTxTypes}}, \code{\link{sumOverGenes}}
}
