% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{assignTxType}
\alias{assignTxType}
\title{Annotate GRanges with transcript type}
\usage{
assignTxType(gr, txdb, tssUpstream = 100, tssDownstream = 100,
  proximalUpstream = 1000)
}
\arguments{
\item{gr}{GRanges: Ranges to annotated.}

\item{txdb}{TxDb: Transcript database to use for annotation.}

\item{tssUpstream}{integer: Distance to extend annotated promoter upstream.}

\item{tssDownstream}{integer: Distance to extend annotated promoter downstream.}

\item{proximalUpstream}{integer: Maximum distance upstream of promoter to be considered proximal.}
}
\value{
character vector of same length as gr containing TxType for each range.
}
\description{
Annotate a set of ranges in a GRanges object with transcript type (TxType) based on their genic context. Transcripts are obtained from a TxDb object.
}
\examples{
# ADD_EXAMPLES_HERE
}

