% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancers.R
\docType{methods}
\name{calcBidirectionality}
\alias{calcBidirectionality}
\alias{calcBidirectionality,GRanges-method}
\alias{calcBidirectionality,RangedSummarizedExperiment-method}
\title{Calculate sample-wise bidirectionally of clusters.}
\usage{
calcBidirectionality(object, ...)

\S4method{calcBidirectionality}{GRanges}(object, samples,
  inputAssay = "counts", outputColumn = "bidirectionality")

\S4method{calcBidirectionality}{RangedSummarizedExperiment}(object, ...)
}
\arguments{
\item{object}{GenomicRanges or RangedSummarizedExperiment: Ranges with pooled CTSSs stored in the score column.}

\item{...}{additional arguments passed to methods.}

\item{samples}{RangedSummarizedExperiment: Sample-wise CTSSs stored as an assay.}

\item{inputAssay}{character: Name of assay in samples holding input CTSS values.}

\item{outputColumn}{character: Name of column in object to hold bidirectionality values.}
}
\value{
object returned with bidirectionality scores added in mcols.
}
\description{
For each bidirectional site, calculates how many individual samples shows transcription in both directions.
}
\examples{
data(exampleCTSSs)
data(exampleBidirectional)

calcBidirectionality(exampleBidirectional, samples=exampleCTSSs)
}
\seealso{
Other Calculation functions: \code{\link{calcComposition}},
  \code{\link{calcPooled}}, \code{\link{calcShape}},
  \code{\link{calcSupport}}, \code{\link{calcTPM}},
  \code{\link{calcTotalTags}},
  \code{\link{subsetByBidirectionality}},
  \code{\link{subsetByComposition}},
  \code{\link{subsetBySupport}}
}
