% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancers.R
\docType{methods}
\name{clusterBidirectionally}
\alias{clusterBidirectionally}
\alias{clusterBidirectionally,GenomicRanges-method}
\alias{clusterBidirectionally,RangedSummarizedExperiment-method}
\alias{clusterBidirectionally,GPos-method}
\title{Bidirectional clustering of pooled CTSSs.}
\usage{
clusterBidirectionally(object, ...)

\S4method{clusterBidirectionally}{GenomicRanges}(object, window = 199,
  balanceThreshold = 0.95)

\S4method{clusterBidirectionally}{RangedSummarizedExperiment}(object, ...)

\S4method{clusterBidirectionally}{GPos}(object, ...)
}
\arguments{
\item{object}{GenomicRanges or RangedSummarizedExperiment: Pooled CTSSs stored in the score column.}

\item{...}{additional arguments passed to methods.}

\item{window}{integer: Width of sliding window used for calculating window sums.}

\item{balanceThreshold}{numeric: Minimum value of the BC to use for slice-reduce, a value of 1 corresponds to perfectly balanced sites.}
}
\value{
GRanges with bidirectional sites: Minimum width is 1 + 2*window, TPM sum (on both strands) in the score column, maximal bidirectional site in the thick column and maximum balance in the balance column.
}
\description{
Finds sites with (balanced and divergent) bidirectional transcription using sliding windows of summed coverage: The Bhattacharyya coefficient (BC) is used to quantify depature from a perfectly balanced site, and a slice-reduce is used to identify sites.
}
\examples{
# ADD_EXAMPLES_HERE
}
