% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\docType{methods}
\name{clusterUnidirectionally}
\alias{clusterUnidirectionally}
\alias{clusterUnidirectionally,GenomicRanges-method}
\alias{clusterUnidirectionally,RangedSummarizedExperiment-method}
\alias{clusterUnidirectionally,GPos-method}
\title{Tag Clustering of pooled CTSSs.}
\usage{
clusterUnidirectionally(object, ...)

\S4method{clusterUnidirectionally}{GenomicRanges}(object, pooledCutoff = 0,
  mergeDist = 20)

\S4method{clusterUnidirectionally}{RangedSummarizedExperiment}(object, ...)

\S4method{clusterUnidirectionally}{GPos}(object, ...)
}
\arguments{
\item{object}{GRanges or RangedSummarizedExperiment: Basepair-wise pooled CTSS.}

\item{...}{additional arguments passed to methods.}

\item{pooledCutoff}{numeric: Minimum pooled value to be considered as TC.}

\item{mergeDist}{integer: Merge TCs within this distance.}
}
\value{
GRanges with TPM sum as the score column, and TC peak as the thick column.
}
\description{
Finds unidirectional Tag Cluster (TCs) with a pooled TPM above a certain threshold using a slice-reduce approach. Addtionally calculates the sum and peak position of the TCs.
}
\seealso{
Other Tag-clustering functions: \code{\link{trimToPeak}},
  \code{\link{tuneTagClustering}}
}
