% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwCoverage.R
\name{coverageOfCTSS}
\alias{coverageOfCTSS}
\title{Calculate global CTSS coverage.}
\usage{
coverageOfCTSS(grl, normalizationFunction = trimAndNormalizeGR,
  biocParallel = bpparam(), ...)
}
\arguments{
\item{grl}{GRangesList or SimpleList: GRanges with CTSSs in the score column.}

\item{normalizationFunction}{function or NULL: Function to preprocess the individal GRanges (See Details), or NULL if data should be considered as already normalized.}

\item{biocParallel}{BiocParallelParam: Settings for parallel backend.}

\item{...}{Additional arguments passed to normalizationFunction.}
}
\value{
GRanges with global summed-coverage as the score column.
}
\description{
Given a series of GRanges, preprocesses and normalizes the CTSS-counts before calculating the global coverage by summing over all samples.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other Coverage functions: \code{\link{supportOfCTSS}},
  \code{\link{trimAndNormalizeGR}}
}
