% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{exampleDesign}
\alias{exampleDesign}
\alias{exampleCTSS}
\alias{exampleCoverage}
\alias{exampleTagClusters}
\alias{exampleBidirectionalClusters}
\title{Example CAGE Data}
\format{Example data from various stages of CAGEfightR:
\describe{
  \item{exampleDesign}{DataFrame: Description of samples, including .bw filenames}
  \item{exampleCTSS}{GRangesList: CTSS data for each sample}
  \item{exampleCoverage}{GRanges: Coverage in TPM across all samples}
  \item{exampleTagCluster}{GRanges: Subset of Tag Clusters}
  \item{exampleBidirectionalCluster}{GRanges: Subset of bidirectional clusters}
}}
\source{
\url{http://pubs.acs.org/doi/abs/10.1021/acsnano.6b07533}
}
\usage{
exampleDesign

exampleCTSS

exampleCoverage

exampleTagClusters

exampleBidirectionalClusters
}
\description{
Subset of the CAGE dataset from the paper "Identification of Gene Transcription Start Sites and Enhancers Responding to Pulmonary Carbon Nanotube Exposure in Vivo".
CTSS data from chr18 and chr19 across 11 mouse (mm9 )samples are included.
}
\keyword{datasets}
