% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{importAsGRangesList}
\alias{importAsGRangesList}
\alias{importAsGRangesList.character}
\alias{importAsGRangesList.default}
\title{Import CTSS data into a GRangesList}
\usage{
importAsGRangesList(ctss, cores = NULL)

\method{importAsGRangesList}{default}(ctss, cores = NULL)

\method{importAsGRangesList}{character}(ctss, cores = NULL)
}
\arguments{
\item{ctss}{character: Vector of paths to CTSS-files.}

\item{cores}{integer: Number of cores to use on import.}
}
\value{
GRangesList of the same length as ctss, containing CTSS scores for each file.
}
\description{
Import a series of files containing CTSS data into a GRangesList. The file extensions are used to determine how the file is loaded, currently supportet is .bed, .bedGraph and .bw.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method (always throws an error)

\item \code{character}: Character method
}}

