% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify.R
\name{quantifyClusters}
\alias{quantifyClusters}
\title{Quantify Tag Clusters (TCs)}
\usage{
quantifyClusters(object, clusters, inputAssay = "counts")
}
\arguments{
\item{object}{RangedSummarizedExperiment: CTSSs.}

\item{clusters}{GRanges: Clusters ro be quantified.}

\item{inputAssay}{character: Name of holding values to be quantified}
}
\value{
RangedSummarizedExperiment with quantified CTSSs for each region in cluster, with seqinfo and colData is copied over.
}
\description{
Quantify Tag Clusters (TCs)
}
