% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify.R
\name{quantifyCTSSs}
\alias{quantifyCTSSs}
\title{Quantify CAGE Transcriptions Start Sites (CTSSs)}
\usage{
quantifyCTSSs(plusStrand, minusStrand, design = NULL, genome = NULL,
  tileWidth = 100000000L)
}
\arguments{
\item{plusStrand}{BigWigFileList: BigWig files with plus-strand CTSS data.}

\item{minusStrand}{BigWigFileList: BigWig files with minus-strand CTSS data.}

\item{design}{DataFrame or data.frame: Additional information on samples.}

\item{genome}{Seqinfo: Genome information. If NULL the smallest common genome will be found using bwCommonGenome.}

\item{tileWidth}{integer: Size of tiles to parallelize over.}
}
\value{
RangedSummarizedExperiment, where assay is a sparse matrix (dgCMatrix) of CTSS counts and rowRanges is GPos-object.
}
\description{
This function reads in CTSS data from a series of BigWig-files and returns a basepair-by-sample count matrix. To save memory, the count matrix is stored as a sparse matrix (dgCMatrix), and basepair positions as GPos-object.
}
