% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantification.R
\name{quantifyFeatures}
\alias{quantifyFeatures}
\alias{quantifyFeatures.GRangesList}
\alias{quantifyFeatures.character}
\alias{quantifyFeatures.default}
\title{Quantify CTSS across features}
\usage{
quantifyFeatures(ctss, features, cores)

\method{quantifyFeatures}{default}(ctss, features, cores)

\method{quantifyFeatures}{GRangesList}(ctss, features, cores = NULL)

\method{quantifyFeatures}{character}(ctss, features, cores = NULL)
}
\arguments{
\item{ctss}{GRangesList or character: CTSS data stores as a GRangesList or a vector of paths to CTSS-files.}

\item{features}{GRanges: Ranges to be quantified.}

\item{cores}{NULL or integer: If not NULL, use multiple cores to quantify features.}
}
\value{
matrix with rows as features and columns as samples.
}
\description{
Counts the number of CTSS-tags in a set of features
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default method (Always throws an error)

\item \code{GRangesList}: high memory method

\item \code{character}: low memory method
}}
\examples{
# ADD_EXAMPLES_HERE
}

