% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwQuantification.R
\name{quantifyFeatures}
\alias{quantifyFeatures}
\title{Quantify CTSS across features}
\usage{
quantifyFeatures(ctss, features, biocParallel = bpparam())
}
\arguments{
\item{ctss}{GRangesList or SimpleList: GRanges with CTSSs in the score column.}

\item{features}{GRanges: Ranges to be quantified.}

\item{biocParallel}{BiocParallelParam: Settings for parallel backend.}
}
\value{
matrix with rows as features and columns as samples.
}
\description{
Counts the number of CTSS-tags in a set of features
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other Quantification functions: \code{\link{countScoredOverlaps}}
}

