% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify.R
\name{quantifyGenes}
\alias{quantifyGenes}
\title{Quantify genes}
\usage{
quantifyGenes(object, genes, inputAssay = "counts")
}
\arguments{
\item{object}{RangedSummarizedExperiment: Transcript-level counts.}

\item{genes}{character: Name of column in rowData holding genes (NAs will be discarded).}

\item{inputAssay}{character: Name of assay holding values to be quantified, usually raw counts.}
}
\value{
RangedSummarizedExperiment with rows corresponding to genes. All other information is copied over from object.
}
\description{
Obtain a gene-level EM by summing transcripts within genes. Unannotated transcripts (NAs) are discarded.
}
