% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSE.R
\name{quickEnhancers}
\alias{quickEnhancers}
\title{Complete pipeline for finding and quantifying enhancers.}
\usage{
quickEnhancers(design, bwPlus = NULL, bwMinus = NULL, ctss = NULL,
  ctssCoverage = NULL, genome = NULL, keepCoverage = FALSE,
  window = 199, balanceThreshold = 0.95, minBidirectionality = 1,
  biocParallel = bpparam())
}
\arguments{
\item{design}{DataFrame or data.frame: Information about individual samples.}

\item{bwPlus}{BigWigFileList: BigWig files with plus-strand CTSS data.}

\item{bwMinus}{BigWigFileList: BigWig files with minus-strand CTSS data.}

\item{ctss}{GRangesList: CTSS data for samples.}

\item{ctssCoverage}{GRanges: CTSS-coverage across all samples.}

\item{genome}{Seqinfo or NULL: If not NULL, use the supplied Seqinfo as genome.}

\item{keepCoverage}{logical: Whether to keep global coverage in the output object.}

\item{window}{integer: Width of sliding window used for calculating sums.}

\item{balanceThreshold}{numeric: Minimum balance score to be considered a bidirectional site.}

\item{minBidirectionality}{integer: Minimum number of samples that show bidirectional transcription.}

\item{biocParallel}{BiocParallelParam: Settings for parallel backend.}
}
\value{
RangedSummarizedExperiment.
}
\description{
Complete pipeline for finding and quantifying enhancers.
}
\examples{
# ADD_EXAMPLES_HERE
}
