% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSE.R
\name{quickTSSs}
\alias{quickTSSs}
\title{Complete pipeline for finding and quantifying TSSs.}
\usage{
quickTSSs(design, bwPlus = NULL, bwMinus = NULL, ctss = NULL,
  ctssCoverage = NULL, genome = NULL, keepCoverage = FALSE,
  mergeDist = 20, stepSize = 10, biocParallel = bpparam())
}
\arguments{
\item{design}{DataFrame or data.frame: Information about individual samples.}

\item{bwPlus}{BigWigFileList: BigWig files with plus-strand CTSS data.}

\item{bwMinus}{BigWigFileList: BigWig files with minus-strand CTSS data.}

\item{ctss}{GRangesList: CTSS data for samples.}

\item{ctssCoverage}{GRanges: CTSS-coverage across all samples.}

\item{genome}{Seqinfo or NULL: If not NULL, use the supplied Seqinfo as genome.}

\item{keepCoverage}{logical: Whether to keep global coverage in the output object.}

\item{mergeDist}{integer: Merge TCs within this distance.}

\item{stepSize}{integer: Size of steps for increasing threshold.}

\item{biocParallel}{BiocParallelParam: Settings for parallel backend.}
}
\value{
RangedSummarizedExperiment.
}
\description{
This functions wraps a number of functions: readCTSS, coverageOfCTSS,
tuneTagclusting, tagClustering & quantifyFeatures. Returns results as a
RangedSummarizedExperiment
}
\examples{
# ADD_EXAMPLES_HERE
}
