% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwIO.R
\name{readCTSS}
\alias{readCTSS}
\title{Import CTSS data from BigWig-files.}
\usage{
readCTSS(bwPlus, bwMinus, genome = NULL, outFormat = "GRangesList",
  biocParallel = bpparam(), ...)
}
\arguments{
\item{bwPlus}{BigWigFileList: BigWig files with plus-strand CTSS data.}

\item{bwMinus}{BigWigFileList: BigWig files with minus-strand CTSS data.}

\item{genome}{Seqinfo or NULL: If not NULL, use the supplied Seqinfo as genome.}

\item{outFormat}{character: Output as either a "GRangesList" or "SimpleList".}

\item{biocParallel}{BiocParallelParam: Settings for parallel backend.}

\item{...}{additional arguments passed to commonBigWigGenome used when attempting to merge genomes..}
}
\value{
Either a GRangesList or a SimpleList holding GRanges with CTSS data as the score column.
}
\description{
This function reads a series of BigWig-files into R as a GRanges. Multiple checks are performed to ensure the genomes of the different BigWig-files are compatible.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other IO functions: \code{\link{checkBigWigGenomes}},
  \code{\link{commonBigWigGenome}},
  \code{\link{fastReadBigWig}},
  \code{\link{uniqueBigWigGenomes}}
}
