% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwCoverage.R
\name{splitByStrand}
\alias{splitByStrand}
\title{Split GRanges by strand.}
\usage{
splitByStrand(gr)
}
\arguments{
\item{gr}{GRanges-object.}
}
\value{
GRangesList with three elements: +, - and *.
}
\description{
Split a GRanges into a GRangesList by strand.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other Helper functions: \code{\link{TPM}}
}

