% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{swapRanges}
\alias{swapRanges}
\title{Swap ranges in a GRanges.}
\usage{
swapRanges(gr, column = "thick")
}
\arguments{
\item{gr}{GRanges: GRanges}

\item{column}{character: Name of column in GRanges containing an IRanges.}
}
\value{
GRanges with swapped ranges. Old ranges are in the column "swap"
}
\description{
Swap out the range of a GRanges-object with another IRanges-object stored inside the same object. I.e., swapping TC widths with TC peaks.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other Helper functions: \code{\link{TPM}},
  \code{\link{extendRanges}},
  \code{\link{resolvedOverlaps}},
  \code{\link{splitByStrand}}
}
