% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_browser.R
\name{tagClusterTrack}
\alias{tagClusterTrack}
\title{Tag Cluster Genome Browser Track}
\usage{
tagClusterTrack(gr, plusColor = "cornflowerblue", minusColor = "tomato",
  ...)
}
\arguments{
\item{gr}{GRanges: GRanges with peaks in the thick-column.}

\item{plusColor}{character: Color for plus-strand coverage.}

\item{minusColor}{character: Color for minus-strand coverage.}

\item{...}{additional names passed on to GeneRegionTrack.}
}
\value{
GeneRegionTrack-object.
}
\description{
Create a Gviz-track of Tag Clusters (TCs), indicating TC peak with thickness and strand with color.
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other Genome Browser functions: \code{\link{strandedCoverageTrack}}
}
