% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimTo.R
\docType{methods}
\name{trimToPercentiles}
\alias{trimToPercentiles}
\alias{trimToPercentiles,GRanges,GenomicRanges-method}
\alias{trimToPercentiles,RangedSummarizedExperiment,GenomicRanges-method}
\alias{trimToPercentiles,GRanges,RangedSummarizedExperiment-method}
\alias{trimToPercentiles,RangedSummarizedExperiment,RangedSummarizedExperiment-method}
\title{Trim width of TCs to expression percentiles}
\usage{
trimToPercentiles(object, pooled, ...)

\S4method{trimToPercentiles}{GRanges,GenomicRanges}(object, pooled,
  percentile = 0.1, symmetric = FALSE)

\S4method{trimToPercentiles}{RangedSummarizedExperiment,GenomicRanges}(object,
  pooled, ...)

\S4method{trimToPercentiles}{GRanges,RangedSummarizedExperiment}(object, pooled,
  ...)


  \S4method{trimToPercentiles}{RangedSummarizedExperiment,RangedSummarizedExperiment}(object,
  pooled, ...)
}
\arguments{
\item{object}{GenomicRanges or RangedSummarizedExperiment: TCs to be trimmed.}

\item{pooled}{GenomicRanges or RangedSummarizedExperiment: CTSS coverage.}

\item{...}{additional arguments passed to methods.}

\item{percentile}{numeric: Fraction of expression to remove from TCs.}

\item{symmetric}{logical: Whether to trim the same amount from both edges of the TC (TRUE) or always trim from the least expressed end (FALSE).}
}
\value{
GRanges with trimmed TCs, including recalculated peaks and scores.
}
\description{
Given a set of TCs and genome-wide CTSS coverage, reduce the width of TC until a certain amount of expression has been removed.
}
\examples{
# ADD_EXAMPLES_HERE
}
